Unify the Metric ids across all the ranks.
I thought I did that before, but oh well... it works now.

+++ src/lib/profile/sinks/packed.cpp
@@ -136,8 +136,8 @@ void Packed::packMetrics(std::vector<std::uint8_t>& out) noexcept {
     pack(out, (std::uint64_t)c.userdata[src.identifier()]);
     for(const Metric& m: metrics) {
       if(auto v = m.getFor(c)) {
-        pack(out, v->get(MetricScope::exclusive).value_or(0));
-        pack(out, v->get(MetricScope::inclusive).value_or(0));
+        pack(out, (double)v->get(MetricScope::exclusive).value_or(0));
+        pack(out, (double)v->get(MetricScope::inclusive).value_or(0));
       } else {
         pack(out, (double)0);
         pack(out, (double)0);
